+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.4
+
+- styling printTitle
+
+## 0.0.3
+
+- Add new method: numberAutoDigit
+
 ## 0.0.2
 
 - Add example